Improves #1456 - Do not increase launch delay
if health check fails.

Detecting health failures takes quite some time. First there is the
grace period while the task is starting up. Then there have to be
maxConsecutiveFailure failures until the task is considered as
unhealthy. If we detect the failure after this, we should not
delay task launching further.